Get `flexctl backup config show` working.
